Added dotenv and jsonwebtoken packages, created user.js router file to register users, edited models
The dotenv and jsonwebtoken packages allow a user to stay logged in once they enter their credentials. The users router
registers users for both students and dining halls. Registration takes in all required fields and encrypts the password.
In addition, it checks that a username and password are not already taken for student registration. Likewise, it checks
that a name is not already taken for registering dining hall users. To use the jsonwebtoken, you have to add a .env file
to the server folder. Within this folder, the variable SECRET_KEY should be set to any string, i.e 'SECRET_KEY'=secret.