Completed the comments, food, likes, users routers
Made minor tweaks to the database creation code to fix some errors
with deletion and table association. Also added a new table called
calcounter, which will be used for the calorie counter feature of
the app. The routers should be working for most of the app's
required functionality. There is some commented code that will need
to be uncommented once the frontend and backend are connected to
support the uploading of images.