Experimental JS backend connection components
Created the POST features to send data to the
backend and implemented frontend endpoints.

Still need to test this implementation as there may
be issues with sending and receiving data from the
right database table based on whether the toggle for
dining hall or student is enabled. There also may
be issues with login.