Updated login and registration pages
Created seperate const functions for dining halls and students in order to differentiate
their data on the backend. THis will be useful for when we connect the frontend and
backend later.

Utilized the state of the toggle button on the login page to decide where to send login
information to be processed by the backend since users and dining halls are handled
seperately on the backend.

Changed the login requirements for dining halls to require the dining hall name
instead of a username.