Main-Page: added liking & made cards dynamic
I also switched the index.js back so that it does not require the weird
    import { Link } from 'react-router-dom';

LikeButton: I implemented a liking system that decrements or increments every time you click the button.
This liking button is not dynamic, but I have suggested a fetch api in the commented code below in LikeButton.js

Cards: I suggested some fetch API code to help connect front and backend that is commented out under CardItem.js
This still needs to be tested, and it does assume that Cards.js will pass in some of the values.