Running front and backend concurrently (not fully working) for login-react
Used cors in index.js, changed the package.json so that we can use 'npm run dev'
It does actually update the database but the console says there are errors.
Changed the URL of the endpoint for backend to prepend "http://localhost:8080".
It does seem to work if we run the backend and frontend on separate terminals as of now.